+++ internal/provider/model_sdwan_transport_management_vpn_profile_parcel.go
@@ -33,24 +33,24 @@ import (
 
 // Section below is generated&owned by "gen/generator.go". //template:begin types
 type TransportManagementVPN struct {
-	Id                                    types.String                             `tfsdk:"id"`
-	Version                               types.Int64                              `tfsdk:"version"`
-	Name                                  types.String                             `tfsdk:"name"`
-	Description                           types.String                             `tfsdk:"description"`
-	FeatureProfileId                      types.String                             `tfsdk:"feature_profile_id"`
-	BasicConfigurationDescription         types.String                             `tfsdk:"basic_configuration_description"`
-	BasicConfigurationDescriptionVariable types.String                             `tfsdk:"basic_configuration_description_variable"`
-	PrimaryDnsAddressIpv4                 types.String                             `tfsdk:"primary_dns_address_ipv4"`
-	PrimaryDnsAddressIpv4Variable         types.String                             `tfsdk:"primary_dns_address_ipv4_variable"`
-	SecondaryDnsAddressIpv4               types.String                             `tfsdk:"secondary_dns_address_ipv4"`
-	SecondaryDnsAddressIpv4Variable       types.String                             `tfsdk:"secondary_dns_address_ipv4_variable"`
-	PrimaryDnsAddressIpv6                 types.String                             `tfsdk:"primary_dns_address_ipv6"`
-	PrimaryDnsAddressIpv6Variable         types.String                             `tfsdk:"primary_dns_address_ipv6_variable"`
-	SecondaryDnsAddressIpv6               types.String                             `tfsdk:"secondary_dns_address_ipv6"`
-	SecondaryDnsAddressIpv6Variable       types.String                             `tfsdk:"secondary_dns_address_ipv6_variable"`
-	NewHostMappings                       []TransportManagementVPNNewHostMappings  `tfsdk:"new_host_mappings"`
-	Ipv4StaticRoutes                      []TransportManagementVPNIpv4StaticRoutes `tfsdk:"ipv4_static_routes"`
-	Ipv6StaticRoutes                      []TransportManagementVPNIpv6StaticRoutes `tfsdk:"ipv6_static_routes"`
+	Id                              types.String                             `tfsdk:"id"`
+	Version                         types.Int64                              `tfsdk:"version"`
+	Name                            types.String                             `tfsdk:"name"`
+	Description                     types.String                             `tfsdk:"description"`
+	FeatureProfileId                types.String                             `tfsdk:"feature_profile_id"`
+	InterfaceDescription            types.String                             `tfsdk:"interface_description"`
+	InterfaceDescriptionVariable    types.String                             `tfsdk:"interface_description_variable"`
+	PrimaryDnsAddressIpv4           types.String                             `tfsdk:"primary_dns_address_ipv4"`
+	PrimaryDnsAddressIpv4Variable   types.String                             `tfsdk:"primary_dns_address_ipv4_variable"`
+	SecondaryDnsAddressIpv4         types.String                             `tfsdk:"secondary_dns_address_ipv4"`
+	SecondaryDnsAddressIpv4Variable types.String                             `tfsdk:"secondary_dns_address_ipv4_variable"`
+	PrimaryDnsAddressIpv6           types.String                             `tfsdk:"primary_dns_address_ipv6"`
+	PrimaryDnsAddressIpv6Variable   types.String                             `tfsdk:"primary_dns_address_ipv6_variable"`
+	SecondaryDnsAddressIpv6         types.String                             `tfsdk:"secondary_dns_address_ipv6"`
+	SecondaryDnsAddressIpv6Variable types.String                             `tfsdk:"secondary_dns_address_ipv6_variable"`
+	NewHostMappings                 []TransportManagementVPNNewHostMappings  `tfsdk:"new_host_mappings"`
+	Ipv4StaticRoutes                []TransportManagementVPNIpv4StaticRoutes `tfsdk:"ipv4_static_routes"`
+	Ipv6StaticRoutes                []TransportManagementVPNIpv6StaticRoutes `tfsdk:"ipv6_static_routes"`
 }
 
 type TransportManagementVPNNewHostMappings struct {
@@ -61,14 +61,14 @@ type TransportManagementVPNNewHostMappings struct {
 }
 
 type TransportManagementVPNIpv4StaticRoutes struct {
-	NetworkAddress                 types.String                                                   `tfsdk:"network_address"`
-	NetworkAddressVariable         types.String                                                   `tfsdk:"network_address_variable"`
-	SubnetMask                     types.String                                                   `tfsdk:"subnet_mask"`
-	SubnetMaskVariable             types.String                                                   `tfsdk:"subnet_mask_variable"`
-	Gateway                        types.String                                                   `tfsdk:"gateway"`
-	Ipv4RouteGatewayNextHo         []TransportManagementVPNIpv4StaticRoutesIpv4RouteGatewayNextHo `tfsdk:"ipv4_route_gateway_next_ho"`
-	AdministrativeDistance         types.Int64                                                    `tfsdk:"administrative_distance"`
-	AdministrativeDistanceVariable types.String                                                   `tfsdk:"administrative_distance_variable"`
+	NetworkAddress                 types.String                                     `tfsdk:"network_address"`
+	NetworkAddressVariable         types.String                                     `tfsdk:"network_address_variable"`
+	SubnetMask                     types.String                                     `tfsdk:"subnet_mask"`
+	SubnetMaskVariable             types.String                                     `tfsdk:"subnet_mask_variable"`
+	Gateway                        types.String                                     `tfsdk:"gateway"`
+	NextHops                       []TransportManagementVPNIpv4StaticRoutesNextHops `tfsdk:"next_hops"`
+	AdministrativeDistance         types.Int64                                      `tfsdk:"administrative_distance"`
+	AdministrativeDistanceVariable types.String                                     `tfsdk:"administrative_distance_variable"`
 }
 
 type TransportManagementVPNIpv6StaticRoutes struct {
@@ -80,7 +80,7 @@ type TransportManagementVPNIpv6StaticRoutes struct {
 	NatVariable    types.String                                     `tfsdk:"nat_variable"`
 }
 
-type TransportManagementVPNIpv4StaticRoutesIpv4RouteGatewayNextHo struct {
+type TransportManagementVPNIpv4StaticRoutesNextHops struct {
 	Address                        types.String `tfsdk:"address"`
 	AddressVariable                types.String `tfsdk:"address_variable"`
 	AdministrativeDistance         types.Int64  `tfsdk:"administrative_distance"`
@@ -119,15 +119,15 @@ func (data TransportManagementVPN) toBody(ctx context.Context) string {
 	body, _ = sjson.Set(body, path+"vpnId.optionType", "default")
 	body, _ = sjson.Set(body, path+"vpnId.value", 512)
 
-	if !data.BasicConfigurationDescriptionVariable.IsNull() {
+	if !data.InterfaceDescriptionVariable.IsNull() {
 		body, _ = sjson.Set(body, path+"name.optionType", "variable")
-		body, _ = sjson.Set(body, path+"name.value", data.BasicConfigurationDescriptionVariable.ValueString())
-	} else if data.BasicConfigurationDescription.IsNull() {
+		body, _ = sjson.Set(body, path+"name.value", data.InterfaceDescriptionVariable.ValueString())
+	} else if data.InterfaceDescription.IsNull() {
 		body, _ = sjson.Set(body, path+"name.optionType", "default")
 
 	} else {
 		body, _ = sjson.Set(body, path+"name.optionType", "global")
-		body, _ = sjson.Set(body, path+"name.value", data.BasicConfigurationDescription.ValueString())
+		body, _ = sjson.Set(body, path+"name.value", data.InterfaceDescription.ValueString())
 	}
 
 	if !data.PrimaryDnsAddressIpv4Variable.IsNull() {
@@ -223,7 +223,7 @@ func (data TransportManagementVPN) toBody(ctx context.Context) string {
 			itemBody, _ = sjson.Set(itemBody, "gateway.value", item.Gateway.ValueString())
 		}
 		itemBody, _ = sjson.Set(itemBody, "nextHop", []interface{}{})
-		for _, childItem := range item.Ipv4RouteGatewayNextHo {
+		for _, childItem := range item.NextHops {
 			itemChildBody := ""
 
 			if !childItem.AddressVariable.IsNull() {
@@ -319,14 +319,14 @@ func (data *TransportManagementVPN) fromBody(ctx context.Context, res gjson.Resu
 		data.Description = types.StringNull()
 	}
 	path := "payload.data."
-	data.BasicConfigurationDescription = types.StringNull()
-	data.BasicConfigurationDescriptionVariable = types.StringNull()
+	data.InterfaceDescription = types.StringNull()
+	data.InterfaceDescriptionVariable = types.StringNull()
 	if t := res.Get(path + "name.optionType"); t.Exists() {
 		va := res.Get(path + "name.value")
 		if t.String() == "variable" {
-			data.BasicConfigurationDescriptionVariable = types.StringValue(va.String())
+			data.InterfaceDescriptionVariable = types.StringValue(va.String())
 		} else if t.String() == "global" {
-			data.BasicConfigurationDescription = types.StringValue(va.String())
+			data.InterfaceDescription = types.StringValue(va.String())
 		}
 	}
 	data.PrimaryDnsAddressIpv4 = types.StringNull()
@@ -430,9 +430,9 @@ func (data *TransportManagementVPN) fromBody(ctx context.Context, res gjson.Resu
 				}
 			}
 			if cValue := v.Get("nextHop"); cValue.Exists() {
-				item.Ipv4RouteGatewayNextHo = make([]TransportManagementVPNIpv4StaticRoutesIpv4RouteGatewayNextHo, 0)
+				item.NextHops = make([]TransportManagementVPNIpv4StaticRoutesNextHops, 0)
 				cValue.ForEach(func(ck, cv gjson.Result) bool {
-					cItem := TransportManagementVPNIpv4StaticRoutesIpv4RouteGatewayNextHo{}
+					cItem := TransportManagementVPNIpv4StaticRoutesNextHops{}
 					cItem.Address = types.StringNull()
 					cItem.AddressVariable = types.StringNull()
 					if t := cv.Get("address.optionType"); t.Exists() {
@@ -453,7 +453,7 @@ func (data *TransportManagementVPN) fromBody(ctx context.Context, res gjson.Resu
 							cItem.AdministrativeDistance = types.Int64Value(va.Int())
 						}
 					}
-					item.Ipv4RouteGatewayNextHo = append(item.Ipv4RouteGatewayNextHo, cItem)
+					item.NextHops = append(item.NextHops, cItem)
 					return true
 				})
 			}
@@ -548,14 +548,14 @@ func (data *TransportManagementVPN) updateFromBody(ctx context.Context, res gjso
 		data.Description = types.StringNull()
 	}
 	path := "payload.data."
-	data.BasicConfigurationDescription = types.StringNull()
-	data.BasicConfigurationDescriptionVariable = types.StringNull()
+	data.InterfaceDescription = types.StringNull()
+	data.InterfaceDescriptionVariable = types.StringNull()
 	if t := res.Get(path + "name.optionType"); t.Exists() {
 		va := res.Get(path + "name.value")
 		if t.String() == "variable" {
-			data.BasicConfigurationDescriptionVariable = types.StringValue(va.String())
+			data.InterfaceDescriptionVariable = types.StringValue(va.String())
 		} else if t.String() == "global" {
-			data.BasicConfigurationDescription = types.StringValue(va.String())
+			data.InterfaceDescription = types.StringValue(va.String())
 		}
 	}
 	data.PrimaryDnsAddressIpv4 = types.StringNull()
@@ -699,10 +699,10 @@ func (data *TransportManagementVPN) updateFromBody(ctx context.Context, res gjso
 				data.Ipv4StaticRoutes[i].Gateway = types.StringValue(va.String())
 			}
 		}
-		for ci := range data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo {
+		for ci := range data.Ipv4StaticRoutes[i].NextHops {
 			keys := [...]string{"address"}
-			keyValues := [...]string{data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].Address.ValueString()}
-			keyValuesVariables := [...]string{data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AddressVariable.ValueString()}
+			keyValues := [...]string{data.Ipv4StaticRoutes[i].NextHops[ci].Address.ValueString()}
+			keyValuesVariables := [...]string{data.Ipv4StaticRoutes[i].NextHops[ci].AddressVariable.ValueString()}
 
 			var cr gjson.Result
 			r.Get("nextHop").ForEach(
@@ -725,24 +725,24 @@ func (data *TransportManagementVPN) updateFromBody(ctx context.Context, res gjso
 					return true
 				},
 			)
-			data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].Address = types.StringNull()
-			data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AddressVariable = types.StringNull()
+			data.Ipv4StaticRoutes[i].NextHops[ci].Address = types.StringNull()
+			data.Ipv4StaticRoutes[i].NextHops[ci].AddressVariable = types.StringNull()
 			if t := cr.Get("address.optionType"); t.Exists() {
 				va := cr.Get("address.value")
 				if t.String() == "variable" {
-					data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AddressVariable = types.StringValue(va.String())
+					data.Ipv4StaticRoutes[i].NextHops[ci].AddressVariable = types.StringValue(va.String())
 				} else if t.String() == "global" {
-					data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].Address = types.StringValue(va.String())
+					data.Ipv4StaticRoutes[i].NextHops[ci].Address = types.StringValue(va.String())
 				}
 			}
-			data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AdministrativeDistance = types.Int64Null()
-			data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AdministrativeDistanceVariable = types.StringNull()
+			data.Ipv4StaticRoutes[i].NextHops[ci].AdministrativeDistance = types.Int64Null()
+			data.Ipv4StaticRoutes[i].NextHops[ci].AdministrativeDistanceVariable = types.StringNull()
 			if t := cr.Get("distance.optionType"); t.Exists() {
 				va := cr.Get("distance.value")
 				if t.String() == "variable" {
-					data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AdministrativeDistanceVariable = types.StringValue(va.String())
+					data.Ipv4StaticRoutes[i].NextHops[ci].AdministrativeDistanceVariable = types.StringValue(va.String())
 				} else if t.String() == "global" {
-					data.Ipv4StaticRoutes[i].Ipv4RouteGatewayNextHo[ci].AdministrativeDistance = types.Int64Value(va.Int())
+					data.Ipv4StaticRoutes[i].NextHops[ci].AdministrativeDistance = types.Int64Value(va.Int())
 				}
 			}
 		}
@@ -868,10 +868,10 @@ func (data *TransportManagementVPN) isNull(ctx context.Context, res gjson.Result
 	if !data.FeatureProfileId.IsNull() {
 		return false
 	}
-	if !data.BasicConfigurationDescription.IsNull() {
+	if !data.InterfaceDescription.IsNull() {
 		return false
 	}
-	if !data.BasicConfigurationDescriptionVariable.IsNull() {
+	if !data.InterfaceDescriptionVariable.IsNull() {
 		return false
 	}
 	if !data.PrimaryDnsAddressIpv4.IsNull() {